requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
aniso8601==9.0.1
bcrypt==3.2.0
blinker==1.4
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
Deprecated==1.2.12
fastai==1.0.61
fastcore==1.3.19
fastprogress==1.0.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-buzz==0.1.15
gunicorn==20.0.4
idna==2.10
image-slicer==2.1.1
importlib-metadata==3.7.3
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
murmurhash==1.0.5
numpy==1.20.1
packaging==20.9
pandas==1.2.3
passlib==1.7.4
pathy==0.4.0
pendulum==2.1.2
Pillow==7.2.0
preshed==3.0.5
py-buzz==1.0.3
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
srsly==2.4.0
thinc==8.0.2
threadpoolctl==2.1.0
torchvision==0.6.0+cpu
tqdm==4.59.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1